espflashesp8266/requirements.txt
# requirements.txt

altgraph==0.17.4
auto-py-to-exe==2.42.0
bitarray==2.9.1
bitstring==4.1.4
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
cryptography==41.0.7
customtkinter==5.2.1
darkdetect==0.8.0
ecdsa==0.18.0
Eel==0.16.0
esptool==4.7.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
intelhex==2.3.0
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pycparser==2.21
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
pyparsing==3.1.1
pyserial==3.5
pywin32-ctypes==0.2.2
PyYAML==6.0.1
reedsolo==1.7.0
setuptools==69.0.3
six==1.16.0
ttkthemes==3.2.2
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.1
